requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
